+++ gazebo/physics/dart/DARTSphereShape.cc
@@ -31,6 +31,7 @@ DARTSphereShape::DARTSphereShape(DARTCollisionPtr _parent)
   : SphereShape(_parent),
     dataPtr(new DARTSphereShapePrivate())
 {
+  _parent->SetDARTCollisionShape(this->dataPtr->dtEllipsoidShape, false);
 }
 
 //////////////////////////////////////////////////
@@ -60,28 +61,7 @@ void DARTSphereShape::SetRadius(double _radius)
 
   SphereShape::SetRadius(_radius);
 
-  DARTCollisionPtr dartCollisionParent =
-      boost::dynamic_pointer_cast<DARTCollision>(this->collisionParent);
-
-  if (dartCollisionParent->GetDARTCollisionShape() == nullptr)
-  {
-    dart::dynamics::BodyNode *dtBodyNode =
-        dartCollisionParent->GetDARTBodyNode();
-    dart::dynamics::EllipsoidShape *dtEllisoidShape =
-        new dart::dynamics::EllipsoidShape(Eigen::Vector3d(_radius*2.0,
-                                                           _radius*2.0,
-                                                           _radius*2.0));
-    dtBodyNode->addCollisionShape(dtEllisoidShape);
-    dartCollisionParent->SetDARTCollisionShape(dtEllisoidShape);
-  }
-  else
-  {
-    dart::dynamics::EllipsoidShape *dtEllipsoidShape =
-        dynamic_cast<dart::dynamics::EllipsoidShape*>(
-          dartCollisionParent->GetDARTCollisionShape());
-    dtEllipsoidShape->setSize(Eigen::Vector3d(_radius*2.0,
-                                              _radius*2.0,
-                                              _radius*2.0));
-  }
+  this->dataPtr->dtEllipsoidShape->setSize(
+        Eigen::Vector3d(_radius*2.0, _radius*2.0, _radius*2.0));
 }